Relocate tournament_results_modified_since from PluginUtils to Utils
