+++ src/utils/__init__.py
@@ -7,12 +7,16 @@ from decimal import Decimal
 from functools import lru_cache, cache
 from math import floor
 from subprocess import CompletedProcess
-from typing import Callable, Iterable, Protocol, Hashable, Collection
+from typing import Callable, Iterable, Protocol, Hashable, Collection, TYPE_CHECKING
 
 import iso4217parse
 import pycountry
 from babel.numbers import format_currency, format_decimal, get_decimal_symbol
 from text_unidecode import unidecode
+
+if TYPE_CHECKING:
+    from data.tournament import Tournament
+    from database.sqlite.event.event_store import StoredTournament
 
 
 class Utils:
@@ -305,6 +309,17 @@ class Utils:
         for property_name in property_names:
             if property_name in obj.__dict__:
                 del obj.__dict__[property_name]
+
+    @staticmethod
+    def tournament_results_modified_since(
+        tournament: 'Tournament | StoredTournament',
+        ref_datetime: datetime,
+    ) -> bool:
+        return ref_datetime < max(
+            tournament.last_update,
+            tournament.last_player_update or datetime.min,
+            tournament.last_pairing_update or datetime.min,
+        )
 
 
 class SupportsEquals(Protocol):
